intermediate moves
