+++ .gitignore
@@ -29,7 +29,7 @@ var/
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
-#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+#  before PyInstaller builds the exe, so as to inject date/others infos into it.
 *.manifest
 *.spec
