+++ README.md
@@ -103,7 +103,7 @@ train(model_q, model_k, device, train_loader, queue, optimizer, epoch)
 ### Скрипт обучения модели на основе Barlow Twins
 
 ```bash
-python src/models/'barlow twins'/barlow_twins.py images_path masks_path target_height target_width batch_size epochs
+python src/models/'barlow_twins'/barlow_twins.py images_path masks_path target_height target_width batch_size epochs
 ```
 
 ```images_path``` - путь к изображениям
