+++ setup.py
@@ -4,7 +4,7 @@ setup(
     name='src',
     packages=find_packages(),
     version='0.1.0',
-    description='Clumped texture composite images projects (froath, rocks and other texture images)',
+    description='Clumped texture composite images projects (froath, rocks and others texture images)',
     author='s70c3',
     license='MIT',
 )
