+++ src/configs/inference/hrnet.yaml
