+++ src/configs/inference/segformer.yaml
