+++ src/configs/inference/swin.yaml
