+++ src/configs/inference/yolov8.yaml
