+++ src/configs/training/data_yolov8_seg.yaml
