+++ src/configs/training/training_hrnet_config.yaml
