+++ src/configs/training/training_swin_config.yaml
