+++ src/configs/training/yolov8_seg.yaml
