+++ src/infrastructure/models_inference/yolov8_seg/yolo_export.py
@@ -4,7 +4,7 @@ Model Export Script
 This script is used for exporting a trained segmentation model to ONNX format and optionally quantizing the model.
 
 The script takes a YAML configuration file as input, which specifies the paths to the trained model checkpoint,
-input tensor shape, and other configuration parameters.
+input tensor shape, and others configuration parameters.
 
 Usage:
     python <model>_export.py <config_path>
@@ -16,7 +16,7 @@ Args:
 
 import sys
 
-from src.models.yolov8_segmentation.model import load_yolov8_segment
+from src.models.yolov8_segmentation.yolov8 import load_yolov8_segment
 from src.models.utils.config import read_yaml_config
