+++ src/others/models/barlow_twins/barlow_twins.py
@@ -1,0 +1,452 @@
+# LICENSE file in the root directory of this source tree.
+
+from pathlib import Path
+import argparse
+import json
+import math
+import os
+import random
+import signal
+import subprocess
+import sys
+import time
+import glob
+
+
+import cv2
+from sklearn.model_selection import train_test_split
+import torch.functional as F
+from PIL import Image, ImageOps, ImageFilter
+from torch import nn, optim
+import torch
+import torchvision
+import torchvision.transforms as transforms
+
+def adjust_learning_rate(args, optimizer, loader, step):
+    max_steps = args.epochs * len(loader)
+    warmup_steps = 10 * len(loader)
+    base_lr = args.batch_size / 256
+    if step < warmup_steps:
+        lr = base_lr * step / warmup_steps
+    else:
+        step -= warmup_steps
+        max_steps -= warmup_steps
+        q = 0.5 * (1 + math.cos(math.pi * step / max_steps))
+        end_lr = base_lr * 0.001
+        lr = base_lr * q + end_lr * (1 - q)
+    optimizer.param_groups[0]['lr'] = lr * args.learning_rate_weights
+    optimizer.param_groups[1]['lr'] = lr * args.learning_rate_biases
+
+
+def handle_sigusr1(signum, frame):
+    os.system(f'scontrol requeue {os.getenv("SLURM_JOB_ID")}')
+    exit()
+
+
+def handle_sigterm(signum, frame):
+    pass
+
+
+def off_diagonal(x):
+    # return a flattened view of the off-diagonal elements of a square matrix
+    n, m = x.shape
+    assert n == m
+    return x.flatten()[:-1].view(n - 1, n + 1)[:, 1:].flatten()
+
+
+class conv_block(nn.Module):
+    def __init__(self, in_c, out_c):
+        super().__init__()
+
+        self.conv1 = nn.Conv2d(in_c, out_c, kernel_size=3, padding=1)
+        self.bn1 = nn.BatchNorm2d(out_c)
+
+        self.conv2 = nn.Conv2d(out_c, out_c, kernel_size=3, padding=1)
+        self.bn2 = nn.BatchNorm2d(out_c)
+
+        self.relu = nn.ReLU()
+
+    def forward(self, inputs):
+        x = self.conv1(inputs)
+        x = self.bn1(x)
+        x = self.relu(x)
+
+        x = self.conv2(x)
+        x = self.bn2(x)
+        x = self.relu(x)
+
+        return x
+
+
+
+
+class encoder_block(nn.Module):
+    def __init__(self, in_c, out_c):
+        super().__init__()
+
+        self.conv = conv_block(in_c, out_c)
+        self.pool = nn.MaxPool2d((2, 2))
+
+    def forward(self, inputs):
+        x = self.conv(inputs)
+        p = self.pool(x)
+
+        return x, p
+
+
+
+class decoder_block(nn.Module):
+    def __init__(self, in_c, out_c):
+        super().__init__()
+
+        self.up = nn.ConvTranspose2d(in_c, out_c, kernel_size=2, stride=2, padding=0)
+        self.conv = conv_block(out_c+out_c, out_c)
+
+    def forward(self, inputs, skip):
+        x = self.up(inputs)
+        x = torch.cat([x, skip], axis=1)
+        x = self.conv(x)
+
+        return x
+
+class unet_enc(nn.Module):
+    def __init__(self):
+        super().__init__()
+        self.e1 = encoder_block(3, 64)
+        self.e2 = encoder_block(64, 128)
+        self.e3 = encoder_block(128, 256)
+        self.e4 = encoder_block(256, 512)
+        self.fc1 = nn.Flatten()
+
+    def forward(self, inputs):
+        self.s1, p1 = self.e1(inputs)
+        self.s2, p2 = self.e2(p1)
+        self.s3, p3 = self.e3(p2)
+        self.s4, p4 = self.e4(p3)
+        fl = self.fc1(p4)
+        return fl
+
+
+class Net(nn.Module):
+    def __init__(self):
+        super(Net, self).__init__()
+        self.conv1 = nn.Conv2d(1, 32, 3, 1)
+        self.conv2 = nn.Conv2d(32, 64, 3, 1)
+        self.fc1 = nn.Linear(9216, 128)
+
+    def forward(self, x):
+        x = self.conv1(x)
+        x = F.relu(x)
+        x = self.conv2(x)
+        x = F.max_pool2d(x, 2)
+        x = torch.flatten(x, 1)
+        x = self.fc1(x)
+        x = F.normalize(x)
+        return x
+class build_unet(nn.Module):
+    def __init__(self):
+        super().__init__()
+
+        self.e1 = encoder_block(3, 64)
+        self.e2 = encoder_block(64, 128)
+        self.e3 = encoder_block(128, 256)
+        self.e4 = encoder_block(256, 512)
+
+        self.b = conv_block(512, 1024)
+
+        self.d1 = decoder_block(1024, 512)
+        self.d2 = decoder_block(512, 256)
+        self.d3 = decoder_block(256, 128)
+        self.d4 = decoder_block(128, 64)
+        self.outputs = nn.Conv2d(64, 1, kernel_size=1, padding=0)
+
+    def forward(self, inputs):
+
+        s1, p1 = self.e1(inputs)
+        s2, p2 = self.e2(p1)
+        s3, p3 = self.e3(p2)
+        s4, p4 = self.e4(p3)
+
+        b = self.b(p4)
+
+
+        d1 = self.d1(b, s4)
+        d2 = self.d2(d1, s3)
+        d3 = self.d3(d2, s2)
+        d4 = self.d4(d3, s1)
+
+
+        outputs = self.outputs(d4)
+
+        return outputs
+
+class BarlowTwins(nn.Module):
+    def __init__(self, args):
+        super().__init__()
+        self.args = args
+        self.backbone = unet_enc()
+        self.backbone.fc = nn.Identity()
+
+        # projector
+        sizes = [2048] + list(map(int, args.projector.split('-')))
+        layers = []
+        for i in range(len(sizes) - 2):
+            layers.append(nn.Linear(sizes[i], sizes[i + 1], bias=False))
+            layers.append(nn.BatchNorm1d(sizes[i + 1]))
+            layers.append(nn.ReLU(inplace=True))
+        layers.append(nn.Linear(sizes[-2], sizes[-1], bias=False))
+        self.projector = nn.Sequential(*layers)
+
+        # normalization layer for the representations z1 and z2
+        self.bn = nn.BatchNorm1d(sizes[-1], affine=False)
+
+    def forward(self, y1, y2):
+        z1 = self.projector(self.backbone(y1))
+        z2 = self.projector(self.backbone(y2))
+
+        # empirical cross-correlation matrix
+        c = self.bn(z1).T @ self.bn(z2)
+
+        # sum the cross-correlation matrix between all gpus
+        c.div_(self.args.batch_size)
+        torch.distributed.all_reduce(c)
+
+        on_diag = torch.diagonal(c).add_(-1).pow_(2).sum()
+        off_diag = off_diagonal(c).pow_(2).sum()
+        loss = on_diag + self.args.lambd * off_diag
+        return loss
+
+
+class LARS(optim.Optimizer):
+    def __init__(self, params, lr, weight_decay=0, momentum=0.9, eta=0.001,
+                 weight_decay_filter=False, lars_adaptation_filter=False):
+        defaults = dict(lr=lr, weight_decay=weight_decay, momentum=momentum,
+                        eta=eta, weight_decay_filter=weight_decay_filter,
+                        lars_adaptation_filter=lars_adaptation_filter)
+        super().__init__(params, defaults)
+
+
+    def exclude_bias_and_norm(self, p):
+        return p.ndim == 1
+
+    @torch.no_grad()
+    def step(self):
+        for g in self.param_groups:
+            for p in g['params']:
+                dp = p.grad
+
+                if dp is None:
+                    continue
+
+                if not g['weight_decay_filter'] or not self.exclude_bias_and_norm(p):
+                    dp = dp.add(p, alpha=g['weight_decay'])
+
+                if not g['lars_adaptation_filter'] or not self.exclude_bias_and_norm(p):
+                    param_norm = torch.norm(p)
+                    update_norm = torch.norm(dp)
+                    one = torch.ones_like(param_norm)
+                    q = torch.where(param_norm > 0.,
+                                    torch.where(update_norm > 0,
+                                                (g['eta'] * param_norm / update_norm), one), one)
+                    dp = dp.mul(q)
+
+                param_state = self.state[p]
+                if 'mu' not in param_state:
+                    param_state['mu'] = torch.zeros_like(p)
+                mu = param_state['mu']
+                mu.mul_(g['momentum']).add_(dp)
+
+                p.add_(mu, alpha=-g['lr'])
+
+
+
+class GaussianBlur(object):
+    def __init__(self, p):
+        self.p = p
+
+    def __call__(self, img):
+        if random.random() < self.p:
+            sigma = random.random() * 1.9 + 0.1
+            return img.filter(ImageFilter.GaussianBlur(sigma))
+        else:
+            return img
+
+
+class Solarization(object):
+    def __init__(self, p):
+        self.p = p
+
+    def __call__(self, img):
+        if random.random() < self.p:
+            return ImageOps.solarize(img)
+        else:
+            return img
+
+
+class Transform:
+    def __init__(self):
+        self.transform = transforms.Compose([
+            transforms.RandomResizedCrop(224, interpolation=Image.BICUBIC),
+            transforms.RandomHorizontalFlip(p=0.5),
+            transforms.RandomApply(
+                [transforms.ColorJitter(brightness=0.4, contrast=0.4,
+                                        saturation=0.2, hue=0.1)],
+                p=0.8
+            ),
+            transforms.RandomGrayscale(p=0.2),
+            GaussianBlur(p=1.0),
+            Solarization(p=0.0),
+            transforms.ToTensor(),
+            transforms.Normalize(mean=[0.485, 0.456, 0.406],
+                                 std=[0.229, 0.224, 0.225])
+        ])
+        self.transform_prime = transforms.Compose([
+            transforms.RandomResizedCrop(224, interpolation=Image.BICUBIC),
+            transforms.RandomHorizontalFlip(p=0.5),
+            transforms.RandomApply(
+                [transforms.ColorJitter(brightness=0.4, contrast=0.4,
+                                        saturation=0.2, hue=0.1)],
+                p=0.8
+            ),
+            transforms.RandomGrayscale(p=0.2),
+            GaussianBlur(p=0.1),
+            Solarization(p=0.2),
+            transforms.ToTensor(),
+            transforms.Normalize(mean=[0.485, 0.456, 0.406],
+                                 std=[0.229, 0.224, 0.225])
+        ])
+
+    def __call__(self, x):
+        y1 = self.transform(x)
+        y2 = self.transform_prime(x)
+        return y1, y2
+
+
+class Dataset:
+    def __init__(self, images_path, masks_path, train_val_split, shape):
+        images_path += "/*.png"
+        masks_path += "/*.png"
+        self.images = [Image.fromarray(cv2.resize(cv2.imread(fname), shape)) for fname in
+                       glob.glob(images_path)]
+        self.masks = [Image.fromarray(cv2.resize(cv2.imread(fname), shape))
+                      for
+                      fname in glob.glob(masks_path)]
+        self.train_val_split = train_val_split
+        self.__split()
+
+    def __split(self):
+        self.train_images, self.val_images, self.train_masks, self.val_masks = train_test_split(self.images, self.masks,
+                                                                                                test_size=self.train_val_split)
+
+    def get_splitted(self):
+        return self.train_images, \
+               self.val_images, \
+               self.train_masks, \
+               self.val_masks
+
+
+if __name__ == '__main__':
+
+    parser = argparse.ArgumentParser(description='')
+
+    parser.add_argument('--data', type=str, help='path to data')
+    parser.add_argument('--height', type=int, help='height of images')
+    parser.add_argument('--width', type=int, help='width of images')
+    parser.add_argument('--size', type=int, help='size of training BT batch')
+    parser.add_argument('--epochs', type=int, help='count of epochs')
+    parser.add_argument('--projector', default='8192-8192-8192', type=str,
+                        metavar='MLP', help='projector MLP')
+
+    print(parser.parse_args().__dict__)
+    args = parser.parse_args()
+
+    images_path = args.images
+    masks_path = args.masks
+    target_height = args.height
+    target_width = args.width
+    batch_size = args.size
+    EPOCHS = args.epochs
+    args = parser.parse_args()
+
+    batchsize = 8
+    epochs = 10
+    out_dir = args.out
+
+    use_cuda = torch.cuda.is_available()
+    device = "cpu"
+
+    kwargs = {'num_workers': 0, 'pin_memory': True}
+
+
+    torch.distributed.init_process_group(
+        backend='nccl', init_method=args.dist_url,
+        world_size=args.world_size, rank=args.rank)
+
+    if args.rank == 0:
+        args.checkpoint_dir.mkdir(parents=True, exist_ok=True)
+        stats_file = open(args.checkpoint_dir / 'stats.txt', 'a', buffering=1)
+        print(' '.join(sys.argv))
+        print(' '.join(sys.argv), file=stats_file)
+
+    torch.backends.cudnn.benchmark = True
+
+    model = BarlowTwins(args)
+    model = nn.SyncBatchNorm.convert_sync_batchnorm(model)
+    param_weights = []
+    param_biases = []
+    for param in model.parameters():
+        if param.ndim == 1:
+            param_biases.append(param)
+        else:
+            param_weights.append(param)
+    parameters = [{'params': param_weights}, {'params': param_biases}]
+    model = torch.nn.parallel.DistributedDataParallel(model)
+    optimizer = LARS(parameters, lr=0, weight_decay=args.weight_decay,
+                     weight_decay_filter=True,
+                     lars_adaptation_filter=True)
+
+    if (args.checkpoint_dir / 'checkpoint.pth').is_file():
+        ckpt = torch.load(args.checkpoint_dir / 'checkpoint.pth',
+                          map_location='cpu')
+        start_epoch = ckpt['epoch']
+        model.load_state_dict(ckpt['model'])
+        optimizer.load_state_dict(ckpt['optimizer'])
+    else:
+        start_epoch = 0
+
+    dataset = torchvision.datasets.ImageFolder(args.data / 'train', Transform())
+    sampler = torch.utils.data.distributed.DistributedSampler(dataset)
+    assert args.batch_size % args.world_size == 0
+    per_device_batch_size = args.batch_size // args.world_size
+    loader = torch.utils.data.DataLoader(
+        dataset, batch_size=per_device_batch_size, num_workers=args.workers,
+        pin_memory=True, sampler=sampler)
+
+    start_time = time.time()
+    scaler = torch.cuda.amp.GradScaler()
+    for epoch in range(start_epoch, args.epochs):
+        sampler.set_epoch(epoch)
+        for step, ((y1, y2), _) in enumerate(loader, start=epoch * len(loader)):
+            adjust_learning_rate(args, optimizer, loader, step)
+            optimizer.zero_grad()
+            with torch.cuda.amp.autocast():
+                loss = model.forward(y1, y2)
+            scaler.scale(loss).backward()
+            scaler.step(optimizer)
+            scaler.update()
+            if step % args.print_freq == 0:
+                if args.rank == 0:
+                    stats = dict(epoch=epoch, step=step,
+                                 lr_weights=optimizer.param_groups[0]['lr'],
+                                 lr_biases=optimizer.param_groups[1]['lr'],
+                                 loss=loss.item(),
+                                 time=int(time.time() - start_time))
+                    print(json.dumps(stats))
+                    print(json.dumps(stats), file=stats_file)
+        if args.rank == 0:
+            state = dict(epoch=epoch + 1, model=model.state_dict(),
+                         optimizer=optimizer.state_dict())
+            torch.save(state, args.checkpoint_dir / 'checkpoint.pth')
+    if args.rank == 0:
+        torch.save(model.module.backbone.state_dict(),
+                   args.checkpoint_dir / 'barlowtwinsUnet.pth')
